[proc] Fire process-exit event for every external process in the job
Previously, this fired the event once per-job, for the job's pgid -
which means the job's lead process.

This was wrong both because

- It fired even for jobs with j->pgid == -2 (e.g. purely internal ones
  that stay in fish's process group)

which could lead to infinite loops when an event-handler listened for
pid 2 and used a builtin.

And

- It fired only for one process even for jobs that have multiple

So for e.g. `command cat | command grep`, only the `cat`s handler was
triggered.

Fixes #4582.