WIP Test the emoji width at compiletime
Since Unicode 9, the width of some characters changed to 2.

Depending on the system, it might have support for it, or it might
not.

Instead of hardcoding specific glibc etc versions, we compile a small
program and check an actual value. This stops the user from having to
define $fish_emoji_width.