set_color: Always print at least a newline
    echo (set_color $somevar)somestring

would otherwise print nothing.

This isn't perfect - `set_color` calls that aren't in command
substitutions, and successive set_color calls in one command
substitutions will now have an extra line.

E.g.

    set_color green; echo -n "banana"

with TERM=dumb, or

    set_color $undefined; echo -n "banana"

with $undefined undefined will now print

    \nbanana

But it's probably much less of an issue - the one case we've hit is in
the `read.expect` tests, which set TERM=dumb, but that's a hack.

Fixes #5443.