screen: Don't decode ENCODE_DIRECT
If e.g. the prompt includes codepoints in the ENCODE_DIRECT area, this
would mangle them, because the prompt never goes through that encoding
process.

Now I'm not sure if this is necessary *at all* in the outputter, but
let's be conservative for now and only do it for anything coming from
the screen.

Might fix #7723