bind: Try to clarify when to use --preset
Sometimes users just look at the output of `bind`, and copy it and
make another `--preset` binding, that will then e.g. be erased when
the preset changes (e.g. after config.fish via
__fish_config_interactive!),
or it will be overridden by a third-party thing.

Do two things to try to remedy this:

1. Print `--user` for user-level bindings
2. If we only print preset bindings, we print a note that explicitly
explains "--preset" should be removed. This note is commented out so
you could still `source` bind's output.

(it's not currently highlighted interactively, I'm thinking of making
a helper for that)

I realize that this is very direct and clumsy, but it's better than
not doing anything - it's explained in the documentation, but few will
read those because they see bind's output and think they understand
it.