Use HOME at runtime, not compile time
When installable, we find our data files at
~/.local/share/fish/install, and for that we need the home directory.

We don't trust $HOME because that can be wrong (especially with
`sudo`), which is also why we don't use a crate or rust's deprecated
std::env::home_dir function.

That's also why $XDG_DATA_HOME isn't a great idea. I'm open to other
directory suggestions.

The code here is pretty ugly, and duplicates some home directory
finding, mostly because our logic for when we get it from getpw* is
weirdly complicated.

But for now it'll do.