Add "flash" bind function to flash
This is something we do regularly in the reader, e.g. when there is no
completion.

This makes it accessible to custom bindings.

E.g. you can e.g. make the
alt-s binding (which toggles a sudo/doas/please prefix) flash instead
of doing anything when the current command is a builtin.