Ask to update if extracted assets are out of date
This checks the version file, and if it exists and is outdated, asks
to update the files.

That means we don't ask if nothing was extracted, so if you just copy
fish to a server where you don't need the man pages, webconfig etc, we
never ask.

But if you *do* extract these, we'll keep them up-to-date.